Add getCurrentGutter function to the JS helper
